Update yargs from 12.0.5 to 16.0.0
We were using a very old version of yargs from 2 years ago.

Dependabot flagged a high severity security issue with y18n, a dependency of yargs.
Version 12.0.5 of yargs required version 4.0.0 of y18n, and the security vulnerability is fixed in 4.0.1 or 5.0.5 of y18n.

Updating from 12.0.5 to 16.0.0 is a big step, but the scripts seem to work as expected.