Add KPI and secondary tag columns to zendesk
Currently, there's an undocumented process in the support recording and reporting where a performance analyst needs to "clean" the data after the CSV has been produced. This adds the necessary columns so that there's no need to do any further data management after the command is run.

This change introduces the KPI column as being hard coded as a blank string. This is a temporary solution to fill this column as there's no way to determine this via the API. The team needs to work out how we manage this going forward.